chore: delete unreferenced symbols left by the workflow removals
Symbol-level follow-up to the module-level sweeps. Each removal was
verified unreferenced across the backend, tests, evals and frontend, then
the scan was re-run to catch what the deletions orphaned in turn.

Backend:
- lib/run_utils.py: maybe_async, call_maybe_async
- lib/workflows/util.py: get_state_by_type_or_raise
- lib/agents/models.py: ValidatedDocument, and DocumentMetadata with it
- lib/models/feedback.py: ClaimPath, and EntityPath with it — Feedback
  addresses entities with a plain dict everywhere, so neither had a caller
- lib/services/references.py: force_match_single_supporting_file and its
  private _get_project_current_revision helper
- lib/services/workflow_runs.py: get_workflow_run_state
- lib/agents/formatting_utils.py: every format_* but format_bibliography,
  the only one still reached (literature_review_v2, live_reports_v2)
- lib/agents/claim_verifier.py: ClaimVerifierAgent and its result models;
  claim_reference_validation_v2 drives its own citation validator, and
  only EvidenceAlignmentLevel and ClaimEvidenceSource outlived it
- lib/agents/literature_review.py: deleted; its one live symbol,
  ReferenceType, moves into methodology_comparator, its sole consumer

Frontend:
- lib/utils.ts: downloadBlobResponse, generateDefaultTestName,
  snakeCaseToTitleCase
- lib/file-download.ts: downloadAsJson
- lib/constants.ts: SUPPORTED_FILE_TYPES
- lib/workflow-state.ts: getChunkErrors — nothing addresses issues by
  chunk since the line-range cutover
- components/assistant-ui/model-selector.tsx: resolveModelEffort
- two barrel files nothing imported (analysis-wizard, results/components);
  results/tabs is still imported by results-visualization and stays
