--- frontend/components/analysis-wizard/index.tsx
@@ -1,4 +1,0 @@
-export { WizardProvider, useWizard, type PreflightStatus, type WizardStep } from './wizard-context';
-export { StepIndicator } from './step-indicator';
-export { StepUpload } from './step-upload';
-export { StepAnalyses } from './step-analyses';
