+++ frontend/components/assistant-ui/model-selector.tsx
@@ -69,19 +69,6 @@ function resolveEffort(
   return efforts?.some((e) => e.id === effort) ? effort : undefined;
 }
 
-/**
- * Returns the effort id if the given model supports it, otherwise undefined.
- * Effort selection is kept sticky across model switches; this resolves what
- * actually applies to the current model.
- */
-export function resolveModelEffort(
-  models: readonly ModelOption[],
-  modelId: string | undefined,
-  effort: string | undefined,
-): string | undefined {
-  return resolveEffort(getModelEfforts(models.find((m) => m.id === modelId)), effort);
-}
-
 function useControllableState<T>({
   prop,
   defaultProp,
