--- frontend/components/results/components/index.ts
@@ -1,6 +1,0 @@
-export {
-  HealthMonitorDashboard,
-  HealthStatusIndicator,
-  OverallHealthCard,
-  WorkflowHealthWidget,
-} from './health-monitor';
