+++ frontend/lib/constants.ts
@@ -1,6 +1,3 @@
 // File upload limits
 export const MAX_FILE_SIZE_MB = 100;
 export const MAX_FILE_SIZE_BYTES = MAX_FILE_SIZE_MB * 1024 * 1024;
-
-// Supported file types for analysis
-export const SUPPORTED_FILE_TYPES = ['.pdf', '.docx', '.doc', '.txt', '.md', '.markdown'];
