+++ frontend/lib/utils.ts
@@ -5,26 +5,12 @@ export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
 
-export function generateDefaultTestName(prefix: string, suffix?: string): string {
-  const timestamp = Date.now();
-  return suffix ? `${prefix}_${suffix}_${timestamp}` : `${prefix}_${timestamp}`;
-}
-
-export async function downloadBlobResponse(apiCall: () => Promise<{ raw: Response }>): Promise<Blob> {
-  const apiResponse = await apiCall();
-  return await apiResponse.raw.blob();
-}
-
 export function formatBytes(bytes: number): string {
   if (bytes === 0) return '0 B';
   const k = 1024;
   const sizes = ['B', 'KB', 'MB', 'GB'];
   const i = Math.floor(Math.log(bytes) / Math.log(k));
   return `${parseFloat((bytes / Math.pow(k, i)).toFixed(1))} ${sizes[i]}`;
-}
-
-export function snakeCaseToTitleCase(s: string): string {
-  return s?.replace(/_/g, ' ').replace(/\b\w/g, (m) => m.toUpperCase()) || '';
 }
 
 /**
