+++ lib/agents/formatting_utils.py
@@ -1,130 +1,4 @@
-from typing import List, Optional
-
-from lib.models.bibliography_item import (
-    BibliographyItem,
-    get_associated_supporting_file,
-)
-from lib.services.file import FileDocument
-from lib.services.vector_store import RetrievedPassage
-from lib.workflows.document_summarization.state import FileSummary
 from lib.workflows.reference_extraction.state import ExtractedReference
-
-
-def format_summary_context(summary: Optional[str]) -> str:
-    """Format summary context for agent prompts."""
-    if not summary:
-        return "No summary available for this document."
-
-    return f"""
-## Summary of the document (for context about the document's main argument)
-```
-{summary}
-```
-"""
-
-
-def format_headings_context(headings: Optional[List[str]]) -> str:
-    """Format headings context for agent prompts."""
-    if not headings:
-        return "No section headings available for this chunk."
-
-    # Format with indentation to show hierarchy
-    formatted = "\n".join(f"{'  ' * i}{heading}" for i, heading in enumerate(headings))
-
-    return f"""
-## Document Section Context
-The following headings provide context about which section of the document this chunk belongs to:
-```
-{formatted}
-```
-"""
-
-
-def format_supporting_documents_prompt_section(
-    supporting_document: FileDocument, truncate_at_character_count: int | None = None
-):
-    markdown = supporting_document.markdown
-    output = f"""
-File name: {supporting_document.file_name}
-File content converted to markdown{" (truncated)" if truncate_at_character_count else ""}:
-```markdown
-{markdown[:truncate_at_character_count] if truncate_at_character_count else markdown}
-```
-"""
-
-    return output
-
-
-def format_supporting_documents_prompt_section_multiple(
-    supporting_files: list[FileDocument],
-    truncate_at_character_count: int | None = None,
-) -> str:
-    supporting_documents = "\n\n".join(
-        [
-            f"""### Supporting document #{index + 1} (index: {index+1})
-{format_supporting_documents_prompt_section(doc, truncate_at_character_count=truncate_at_character_count)}
-"""
-            for index, doc in enumerate(supporting_files or [])
-        ]
-    )
-    return supporting_documents
-
-
-def format_bibliography_item_prompt_section(
-    index: int,
-    item: BibliographyItem,
-    supporting_files: List[FileDocument],
-    summaries: Optional[List["FileSummary"]] = None,
-) -> str:
-    result = f"""### Bibliography entry #{index + 1}
-{item.text}"""
-
-    # If this bibliography item has an associated supporting document, include its summary
-    supporting_file = get_associated_supporting_file(item, supporting_files)
-    if supporting_file and summaries:
-        # Find the summary for this file by file_id
-        summary = next(
-            (s for s in summaries if s.file_id == supporting_file.file_id),
-            None,
-        )
-        if summary:
-            result += f"""
-
-#### Summary of the associated document
-{summary.summary}
-
-"""
-
-    return result
-
-
-def format_bibliography_prompt_section(
-    references: List[BibliographyItem],
-    supporting_files: List[FileDocument],
-    summaries: Optional[List["FileSummary"]] = None,
-) -> str:
-    return "\n\n".join(
-        [
-            format_bibliography_item_prompt_section(
-                index, item, supporting_files, summaries
-            )
-            for index, item in enumerate(references)
-        ]
-    )
-
-
-def format_retrieved_passages(passages: List["RetrievedPassage"]) -> str:
-    """Format retrieved passages from RAG for use as cited references."""
-    if not passages:
-        return "No relevant passages found in supporting documents."
-
-    formatted = []
-    for i, passage in enumerate(passages, 1):
-        formatted.append(
-            f"#### Retrieved Passage {i} from {passage.source_file} (cosine distance: {passage.cosine_distance:.2f})\n"
-            f"```\n{passage.content}\n```\n"
-        )
-    return "\n".join(formatted)
 
 
 def format_bibliography(references: list[ExtractedReference]) -> str:
