+++ lib/run_utils.py
@@ -106,26 +106,6 @@ async def run_tasks(
     return task_results, task_errors
 
 
-def maybe_async(func):
-    """Decorator that makes any function callable with await, regardless of sync/async."""
-
-    async def wrapper(*args, **kwargs):
-        if asyncio.iscoroutinefunction(func):
-            return await func(*args, **kwargs)
-        else:
-            return func(*args, **kwargs)
-
-    return wrapper
-
-
-async def call_maybe_async(func, *args, **kwargs):
-    """Call a function, handling both sync and async cases automatically."""
-    if asyncio.iscoroutinefunction(func):
-        return await func(*args, **kwargs)
-    else:
-        return func(*args, **kwargs)
-
-
 def convert_exceptions_to_workflow_errors(
     task_name: str,
     exceptions: List[Exception | None],
